server/requirements.txt
Flask
Flask-CORS
google-cloud-bigquery
google-generativeai